+++ keycluster/src/module.h
@@ -18,25 +18,18 @@
     #define MODULE_KEY_COUNT KEYBOARD_VECTOR_ITEMS_NUM
     #define MODULE_POINTER_COUNT 1
 
-    #define TEST_LED_GPIO  GPIOA
-    #define TEST_LED_PORT  PORTA
-    #define TEST_LED_CLOCK kCLOCK_PortA
-    #define TEST_LED_PIN   18
+    #define TEST_LED_GPIO  GPIOB
+    #define TEST_LED_PORT  PORTB
+    #define TEST_LED_CLOCK kCLOCK_PortB
+    #define TEST_LED_PIN   11
+
+    #define SDB_GPIO  GPIOA
+    #define SDB_PORT  PORTA
+    #define SDB_CLOCK kCLOCK_PortA
+    #define SDB_PIN   5
 
     #define KEY_ARRAY_TYPE KEY_ARRAY_TYPE_VECTOR
     #define KEYBOARD_VECTOR_ITEMS_NUM 6
-
-// Typedefs
-
-    typedef struct {
-        clock_ip_name_t clock;
-        PORT_Type *port;
-        uint32_t pin;
-        port_mux_t mux;
-        TPM_Type *tpmBase;
-        tpm_chnl_t chnlNumber;
-        uint8_t dutyCyclePercent;
-    } tpm_channel_t;
 
 // Variables: